morudall/requirements.txt
paho-mqtt==1.1
pymongo==3.0.3
python-dateutil==2.4.2
redis==2.10.3
